+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Add status submit message in all languages.
 - Allow uploading the same file for different boreholes.
 - Display an error message if a file is already uploaded for a particular borehole.
+- Resolve error when selecting `Instrumentation` with no EDIT role and no stratigraphy defined for borehole.
 
 ## v2.0.186 - 2023-02-13